WIP on fixing the reload plugins command so nvdaHelperRemote functionality (including buffers) and config profile triggers don't break.
This *almost* works. Unfortunately, the one remaining problem is that the appModule property on any NVDAObjects still alive becomes None, since the old AppModule is dead. There's no way to access all NVDAObjects to tell them to refetch their AppModule. We can't just always refetch if the old AppModule is dead, since we can't assume there won't be another process with the same id later; we should only do this for reload.